Store Refactor - Part 2
