+++ internal/core/interfaces.go
@@ -7,6 +7,6 @@ import (
 // SyncAdapter is used for synchronizing store state to
 // an external system
 type SyncAdapter interface {
-	Sync(ctx context.Context, gateway ResolvedGateway) error
+	Sync(ctx context.Context, gateway ResolvedGateway) (bool, error)
 	Clear(ctx context.Context, id GatewayID) error
 }
