+++ internal/k8s/reconciler/route_test.go
@@ -14,8 +14,6 @@ import (
 
 	clientMocks "github.com/hashicorp/consul-api-gateway/internal/k8s/gatewayclient/mocks"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/reconciler/state"
-	"github.com/hashicorp/consul-api-gateway/internal/k8s/service"
-	"github.com/hashicorp/consul-api-gateway/internal/k8s/service/mocks"
 )
 
 func TestRouteID(t *testing.T) {
@@ -142,180 +140,17 @@ func TestRouteMatchesHostname(t *testing.T) {
 		Spec: gw.HTTPRouteSpec{
 			Hostnames: []gw.Hostname{"*"},
 		},
-	}, state.NewRouteState()).MatchesHostname(&hostname))
+	}, state.NewRouteState()).matchesHostname(&hostname))
 
 	require.False(t, factory.NewRoute(&gw.HTTPRoute{
 		Spec: gw.HTTPRouteSpec{
 			Hostnames: []gw.Hostname{"other.text"},
 		},
-	}, state.NewRouteState()).MatchesHostname(&hostname))
+	}, state.NewRouteState()).matchesHostname(&hostname))
 
 	// check where the underlying route doesn't implement
 	// a matching routine
-	require.True(t, factory.NewRoute(&gw.TCPRoute{}, state.NewRouteState()).MatchesHostname(&hostname))
-}
-
-func TestRouteValidate(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	resolver := mocks.NewMockBackendResolver(ctrl)
-
-	factory := NewFactory(FactoryConfig{
-		Logger:   hclog.NewNullLogger(),
-		Resolver: resolver,
-	})
-
-	require.NoError(t, factory.NewRoute(&core.Pod{}, state.NewRouteState()).Validate(context.Background()))
-
-	require.True(t, factory.NewRoute(&gw.HTTPRoute{}, state.NewRouteState()).IsValid())
-
-	reference := gw.BackendObjectReference{
-		Name: "expected",
-	}
-	resolved := &service.ResolvedReference{
-		Type:      service.ConsulServiceReference,
-		Reference: &service.BackendReference{},
-	}
-
-	resolver.EXPECT().Resolve(gomock.Any(), gomock.Any(), reference).Return(resolved, nil)
-
-	route := factory.NewRoute(&gw.HTTPRoute{
-		Spec: gw.HTTPRouteSpec{
-			Rules: []gw.HTTPRouteRule{{
-				BackendRefs: []gw.HTTPBackendRef{{
-					BackendRef: gw.BackendRef{
-						BackendObjectReference: reference,
-					},
-				}},
-			}},
-		},
-	}, state.NewRouteState())
-	require.NoError(t, route.Validate(context.Background()))
-	require.True(t, route.IsValid())
-
-	expected := errors.New("expected")
-	resolver.EXPECT().Resolve(gomock.Any(), gomock.Any(), reference).Return(nil, expected)
-	require.Equal(t, expected, route.Validate(context.Background()))
-
-	resolver.EXPECT().Resolve(gomock.Any(), gomock.Any(), reference).Return(nil, service.NewK8sResolutionError("error"))
-	require.NoError(t, route.Validate(context.Background()))
-	require.False(t, route.IsValid())
-}
-
-func TestRouteValidateDontAllowCrossNamespace(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	resolver := mocks.NewMockBackendResolver(ctrl)
-	client := clientMocks.NewMockClient(ctrl)
-
-	factory := NewFactory(FactoryConfig{
-		Logger:   hclog.NewNullLogger(),
-		Resolver: resolver,
-		Client:   client,
-	})
-
-	//set up backend ref with a different namespace
-	namespace := gw.Namespace("test")
-	route := factory.NewRoute(&gw.HTTPRoute{
-		Spec: gw.HTTPRouteSpec{
-			Rules: []gw.HTTPRouteRule{{
-				BackendRefs: []gw.HTTPBackendRef{{
-					BackendRef: gw.BackendRef{
-						BackendObjectReference: gw.BackendObjectReference{
-							Name:      "expected",
-							Namespace: &namespace,
-						},
-					},
-				}},
-			}},
-		},
-	}, state.NewRouteState())
-
-	client.EXPECT().
-		GetReferencePoliciesInNamespace(gomock.Any(), gomock.Any()).
-		Return([]gw.ReferencePolicy{
-			{
-				Spec: gw.ReferencePolicySpec{
-					From: []gw.ReferencePolicyFrom{},
-					To:   []gw.ReferencePolicyTo{},
-				},
-			},
-		}, nil)
-
-	// FUTURE Assert appropriate status set on route and !route.IsValid() once ReferencePolicy requirement is enforced
-	_ = route.Validate(context.Background())
-}
-
-// TestRouteValidateAllowCrossNamespaceWithReferencePolicy verifies that a cross-namespace
-// route + backend combination is allowed if an applicable ReferencePolicy is found.
-func TestRouteValidateAllowCrossNamespaceWithReferencePolicy(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	resolver := mocks.NewMockBackendResolver(ctrl)
-	client := clientMocks.NewMockClient(ctrl)
-
-	factory := NewFactory(FactoryConfig{
-		Logger:   hclog.NewNullLogger(),
-		Resolver: resolver,
-		Client:   client,
-	})
-
-	//set up backend ref with a different namespace
-	backendGroup := gw.Group("")
-	backendKind := gw.Kind("Service")
-	backendNamespace := gw.Namespace("namespace2")
-	backendName := gw.ObjectName("backend2")
-	route := factory.NewRoute(&gw.HTTPRoute{
-		ObjectMeta: meta.ObjectMeta{Namespace: "namespace1"},
-		TypeMeta:   meta.TypeMeta{APIVersion: "gateway.networking.k8s.io/v1alpha2", Kind: "HTTPRoute"},
-		Spec: gw.HTTPRouteSpec{
-			Rules: []gw.HTTPRouteRule{{
-				BackendRefs: []gw.HTTPBackendRef{{
-					BackendRef: gw.BackendRef{
-						BackendObjectReference: gw.BackendObjectReference{
-							Group:     &backendGroup,
-							Kind:      &backendKind,
-							Name:      backendName,
-							Namespace: &backendNamespace,
-						},
-					},
-				}},
-			}},
-		},
-	}, state.NewRouteState())
-
-	referencePolicy := gw.ReferencePolicy{
-		TypeMeta:   meta.TypeMeta{},
-		ObjectMeta: meta.ObjectMeta{Namespace: "namespace2"},
-		Spec: gw.ReferencePolicySpec{
-			From: []gw.ReferencePolicyFrom{{
-				Group:     "gateway.networking.k8s.io",
-				Kind:      "HTTPRoute",
-				Namespace: "namespace1",
-			}},
-			To: []gw.ReferencePolicyTo{{
-				Group: "",
-				Kind:  "Service",
-				Name:  &backendName,
-			}},
-		},
-	}
-
-	client.EXPECT().
-		GetReferencePoliciesInNamespace(gomock.Any(), gomock.Any()).
-		Return([]gw.ReferencePolicy{referencePolicy}, nil)
-
-	resolver.EXPECT().
-		Resolve(gomock.Any(), gomock.Any(), gomock.Any()).
-		Return(&service.ResolvedReference{Type: service.ConsulServiceReference, Reference: &service.BackendReference{}}, nil)
-
-	require.NoError(t, route.Validate(context.Background()))
+	require.True(t, factory.NewRoute(&gw.TCPRoute{}, state.NewRouteState()).matchesHostname(&hostname))
 }
 
 func TestRouteResolve(t *testing.T) {
@@ -332,15 +167,9 @@ func TestRouteResolve(t *testing.T) {
 	}
 	listener := gw.Listener{}
 
-	require.Nil(t, factory.NewRoute(&gw.HTTPRoute{}, state.NewRouteState()).Resolve(nil))
+	require.Nil(t, factory.NewRoute(&core.Pod{}, state.NewRouteState()).resolve("", gateway, listener))
 
-	require.Nil(t, factory.NewRoute(&core.Pod{}, state.NewRouteState()).Resolve(NewK8sListener(gateway, listener, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})))
-
-	require.NotNil(t, factory.NewRoute(&gw.HTTPRoute{}, state.NewRouteState()).Resolve(NewK8sListener(gateway, listener, K8sListenerConfig{
-		Logger: hclog.NewNullLogger(),
-	})))
+	require.NotNil(t, factory.NewRoute(&gw.HTTPRoute{}, state.NewRouteState()).resolve("", gateway, listener))
 }
 
 func TestRouteSyncStatus(t *testing.T) {
@@ -355,15 +184,6 @@ func TestRouteSyncStatus(t *testing.T) {
 		Client: client,
 	})
 
-	g := &gw.Gateway{
-		ObjectMeta: meta.ObjectMeta{
-			Name: "expected",
-		},
-	}
-	gateway := factory.NewGateway(NewGatewayConfig{
-		Gateway: g,
-		State:   state.InitialGatewayState(g),
-	})
 	inner := &gw.TCPRoute{
 		Spec: gw.TCPRouteSpec{
 			CommonRouteSpec: gw.CommonRouteSpec{
@@ -396,11 +216,41 @@ func TestRouteSyncStatus(t *testing.T) {
 		},
 	}
 	route := factory.NewRoute(inner, state.NewRouteState())
-	route.OnBound(gateway)
+	route.RouteState.Bound(gw.ParentRef{
+		Name: "expected",
+	})
 
 	expected := errors.New("expected")
 	client.EXPECT().UpdateStatus(gomock.Any(), inner).Return(expected)
 	require.True(t, errors.Is(route.SyncStatus(context.Background()), expected))
 
 	require.NoError(t, route.SyncStatus(context.Background()))
+}
+
+func TestRouteMatchesListenerHostname(t *testing.T) {
+	t.Parallel()
+
+	hostname := gw.Hostname("name")
+	require.True(t, routeMatchesListenerHostname(nil, nil))
+	require.True(t, routeMatchesListenerHostname(&hostname, nil))
+	require.True(t, routeMatchesListenerHostname(&hostname, []gw.Hostname{"*"}))
+	require.False(t, routeMatchesListenerHostname(&hostname, []gw.Hostname{"other"}))
+}
+
+func TestHostnamesMatch(t *testing.T) {
+	t.Parallel()
+
+	require.True(t, hostnamesMatch("*", "*"))
+	require.True(t, hostnamesMatch("", "*"))
+	require.True(t, hostnamesMatch("*", ""))
+	require.True(t, hostnamesMatch("", ""))
+	require.True(t, hostnamesMatch("*.test", "*.test"))
+	require.True(t, hostnamesMatch("a.test", "*.test"))
+	require.True(t, hostnamesMatch("*.test", "a.test"))
+	require.False(t, hostnamesMatch("*.test", "a.b.test"))
+	require.False(t, hostnamesMatch("a.b.test", "*.test"))
+	require.True(t, hostnamesMatch("a.b.test", "*.b.test"))
+	require.True(t, hostnamesMatch("*.b.test", "a.b.test"))
+	require.False(t, hostnamesMatch("*.b.test", "a.c.test"))
+	require.True(t, hostnamesMatch("a.b.test", "a.b.test"))
 }
