+++ internal/k8s/reconciler/state/route.go
@@ -1,8 +1,10 @@
 package state
 
 import (
+	"github.com/hashicorp/consul-api-gateway/internal/k8s/reconciler/common"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/reconciler/status"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/service"
+	gw "sigs.k8s.io/gateway-api/apis/v1alpha2"
 )
 
 // RouteState holds ephemeral state for routes
@@ -19,4 +21,16 @@ func NewRouteState() *RouteState {
 		ResolutionErrors: service.NewResolutionErrors(),
 		ParentStatuses:   make(status.RouteStatuses),
 	}
+}
+
+func (r *RouteState) BindFailed(err error, ref gw.ParentRef) {
+	r.ParentStatuses.BindFailed(r.ResolutionErrors, err, common.AsJSON(ref))
+}
+
+func (r *RouteState) Bound(ref gw.ParentRef) {
+	r.ParentStatuses.Bound(common.AsJSON(ref))
+}
+
+func (r *RouteState) Remove(ref gw.ParentRef) {
+	r.ParentStatuses.Remove(common.AsJSON(ref))
 }
