--- internal/k8s/reconciler/utils.go
@@ -1,262 +1,0 @@
-package reconciler
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"reflect"
-	"strings"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	klabels "k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
-
-	gw "sigs.k8s.io/gateway-api/apis/v1alpha2"
-
-	"github.com/hashicorp/consul-api-gateway/internal/k8s/gatewayclient"
-)
-
-const (
-	// NamespaceNameLabel represents that label added automatically to namespaces is newer Kubernetes clusters
-	NamespaceNameLabel = "kubernetes.io/metadata.name"
-)
-
-func routeMatchesListener(listenerName gw.SectionName, sectionName *gw.SectionName) (bool, bool) {
-	if sectionName == nil {
-		return true, false
-	}
-	return listenerName == *sectionName, true
-}
-
-func routeMatchesListenerHostname(listenerHostname *gw.Hostname, hostnames []gw.Hostname) bool {
-	if listenerHostname == nil || len(hostnames) == 0 {
-		return true
-	}
-
-	for _, name := range hostnames {
-		if hostnamesMatch(name, *listenerHostname) {
-			return true
-		}
-	}
-	return false
-}
-
-func hostnamesMatch(a, b gw.Hostname) bool {
-	if a == "" || a == "*" || b == "" || b == "*" {
-		// any wildcard always matches
-		return true
-	}
-
-	if strings.HasPrefix(string(a), "*.") || strings.HasPrefix(string(b), "*.") {
-		aLabels, bLabels := strings.Split(string(a), "."), strings.Split(string(b), ".")
-		if len(aLabels) != len(bLabels) {
-			return false
-		}
-
-		for i := 1; i < len(aLabels); i++ {
-			if !strings.EqualFold(aLabels[i], bLabels[i]) {
-				return false
-			}
-		}
-		return true
-	}
-
-	return a == b
-}
-
-func routeKindIsAllowedForListener(kinds []gw.RouteGroupKind, route *K8sRoute) bool {
-	if kinds == nil {
-		return true
-	}
-
-	gvk := route.GroupVersionKind()
-	for _, kind := range kinds {
-		group := gw.GroupName
-		if kind.Group != nil && *kind.Group != "" {
-			group = string(*kind.Group)
-		}
-		if string(kind.Kind) == gvk.Kind && group == gvk.Group {
-			return true
-		}
-	}
-
-	return false
-}
-
-// routeAllowedForListenerNamespaces determines whether the route is allowed
-// to bind to the Gateway based on the AllowedRoutes namespace selectors.
-func routeAllowedForListenerNamespaces(ctx context.Context, gatewayNS string, allowedRoutes *gw.AllowedRoutes, route *K8sRoute, c gatewayclient.Client) (bool, error) {
-	var namespaceSelector *gw.RouteNamespaces
-	if allowedRoutes != nil {
-		// check gateway namespace
-		namespaceSelector = allowedRoutes.Namespaces
-	}
-
-	// set default if namespace selector is nil
-	from := gw.NamespacesFromSame
-	if namespaceSelector != nil && namespaceSelector.From != nil && *namespaceSelector.From != "" {
-		from = *namespaceSelector.From
-	}
-	switch from {
-	case gw.NamespacesFromAll:
-		return true, nil
-	case gw.NamespacesFromSame:
-		return gatewayNS == route.GetNamespace(), nil
-	case gw.NamespacesFromSelector:
-		namespaceSelector, err := metav1.LabelSelectorAsSelector(namespaceSelector.Selector)
-		if err != nil {
-			return false, fmt.Errorf("error parsing label selector: %w", err)
-		}
-
-		// retrieve the route's namespace and determine whether selector matches
-		namespace, err := c.GetNamespace(ctx, types.NamespacedName{Name: route.GetNamespace()})
-		if err != nil {
-			return false, fmt.Errorf("error retrieving namespace for route: %w", err)
-		}
-
-		return namespaceSelector.Matches(toNamespaceSet(namespace.GetName(), namespace.GetLabels())), nil
-	}
-	return false, nil
-}
-
-// routeAllowedForBackendRef determines whether the route is allowed
-// for the backend either by being in the same namespace or by having
-// an applicable ReferencePolicy in the same namespace as the backend.
-//
-// TODO This func is currently called once for each backendRef on a route and results
-//   in fetching ReferencePolicies more than we technically have to in some cases
-func routeAllowedForBackendRef(ctx context.Context, route Route, backendRef gw.BackendRef, c gatewayclient.Client) (bool, error) {
-	backendNamespace := ""
-	if backendRef.Namespace != nil {
-		backendNamespace = string(*backendRef.Namespace)
-	}
-
-	// Allow if route and backend are in the same namespace
-	if backendNamespace == "" || route.GetNamespace() == backendNamespace {
-		return true, nil
-	}
-
-	// Allow if ReferencePolicy present for route + backend combination
-	refPolicies, err := c.GetReferencePoliciesInNamespace(ctx, backendNamespace)
-	if err != nil || len(refPolicies) == 0 {
-		return false, err
-	}
-
-	for _, refPolicy := range refPolicies {
-		// Check for a From that applies to the route
-		validFrom := false
-		for _, from := range refPolicy.Spec.From {
-			// If this policy allows the group, kind and namespace for this route
-			if route.GroupVersionKind().Group == string(from.Group) &&
-				route.GroupVersionKind().Kind == string(from.Kind) &&
-				route.GetNamespace() == string(from.Namespace) {
-				validFrom = true
-				break
-			}
-		}
-
-		// If this ReferencePolicy has no applicable From, no need to check for a To
-		if !validFrom {
-			continue
-		}
-
-		// Backend group should default to empty string if not set
-		backendRefGroup := gw.Group("")
-		if backendRef.Group != nil {
-			backendRefGroup = *backendRef.Group
-		}
-
-		// Backend kind should default to Service if not set
-		// TODO Should we default to Service here or go look up the kind from K8s API
-		//   See https://github.com/kubernetes-sigs/gateway-api/issues/1092
-		backendRefKind := gw.Kind("Service")
-		if backendRef.Kind != nil {
-			backendRefKind = *backendRef.Kind
-		}
-
-		// Check for a To that applies to the backendRef
-		for _, to := range refPolicy.Spec.To {
-			// If this policy allows the group, kind, and name for this backend
-			if to.Group == backendRefGroup &&
-				to.Kind == backendRefKind &&
-				(to.Name == nil || *to.Name == backendRef.Name) {
-				return true, nil
-			}
-		}
-	}
-
-	return false, err
-}
-
-func toNamespaceSet(name string, labels map[string]string) klabels.Labels {
-	// If namespace label is not set, implicitly insert it to support older Kubernetes versions
-	if labels[NamespaceNameLabel] == name {
-		// Already set, avoid copies
-		return klabels.Set(labels)
-	}
-	// First we need a copy to not modify the underlying object
-	ret := make(map[string]string, len(labels)+1)
-	for k, v := range labels {
-		ret[k] = v
-	}
-	ret[NamespaceNameLabel] = name
-	return klabels.Set(ret)
-}
-
-func asJSON(item interface{}) string {
-	data, err := json.Marshal(item)
-	if err != nil {
-		// everything passed to this internally should be
-		// serializable, if something is passed to it that
-		// isn't, just panic since it's a usage error at
-		// that point
-		panic(err)
-	}
-	return string(data)
-}
-
-func conditionEqual(a, b metav1.Condition) bool {
-	if a.Type != b.Type ||
-		a.ObservedGeneration != b.ObservedGeneration ||
-		a.Status != b.Status ||
-		a.Reason != b.Reason ||
-		a.Message != b.Message {
-		return false
-	}
-	return true
-}
-
-func conditionsEqual(a, b []metav1.Condition) bool {
-	if len(a) != len(b) {
-		// we have a different number of conditions, so they aren't the same
-		return false
-	}
-
-	for i, newCondition := range a {
-		if !conditionEqual(newCondition, b[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-func listenerStatusEqual(a, b gw.ListenerStatus) bool {
-	if a.Name != b.Name {
-		return false
-	}
-	if !reflect.DeepEqual(a.SupportedKinds, b.SupportedKinds) {
-		return false
-	}
-	if a.AttachedRoutes != b.AttachedRoutes {
-		return false
-	}
-	return conditionsEqual(a.Conditions, b.Conditions)
-}
-
-func getServiceID(name gw.ObjectName, namespace *gw.Namespace, parentNamespace string) string {
-	serviceID := fmt.Sprintf("%s/%s", name, parentNamespace)
-	if namespace != nil {
-		serviceID = fmt.Sprintf("%s/%s", name, *namespace)
-	}
-	return serviceID
-}
