+++ internal/store/mocks/interfaces.go
@@ -36,6 +36,35 @@ func (m *MockStatusTrackingGateway) EXPECT() *MockStatusTrackingGatewayMockRecor
 	return m.recorder
 }
 
+// Bind mocks base method.
+func (m *MockStatusTrackingGateway) Bind(ctx context.Context, route store.Route) []string {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "Bind", ctx, route)
+	ret0, _ := ret[0].([]string)
+	return ret0
+}
+
+// Bind indicates an expected call of Bind.
+func (mr *MockStatusTrackingGatewayMockRecorder) Bind(ctx, route interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Bind", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Bind), ctx, route)
+}
+
+// CanFetchSecrets mocks base method.
+func (m *MockStatusTrackingGateway) CanFetchSecrets(ctx context.Context, secrets []string) (bool, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "CanFetchSecrets", ctx, secrets)
+	ret0, _ := ret[0].(bool)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// CanFetchSecrets indicates an expected call of CanFetchSecrets.
+func (mr *MockStatusTrackingGatewayMockRecorder) CanFetchSecrets(ctx, secrets interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanFetchSecrets", reflect.TypeOf((*MockStatusTrackingGateway)(nil).CanFetchSecrets), ctx, secrets)
+}
+
 // ID mocks base method.
 func (m *MockStatusTrackingGateway) ID() core.GatewayID {
 	m.ctrl.T.Helper()
@@ -50,60 +79,32 @@ func (mr *MockStatusTrackingGatewayMockRecorder) ID() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockStatusTrackingGateway)(nil).ID))
 }
 
-// Listeners mocks base method.
-func (m *MockStatusTrackingGateway) Listeners() []store.Listener {
+// Remove mocks base method.
+func (m *MockStatusTrackingGateway) Remove(ctx context.Context, id string) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Listeners")
-	ret0, _ := ret[0].([]store.Listener)
+	ret := m.ctrl.Call(m, "Remove", ctx, id)
+	ret0, _ := ret[0].(error)
 	return ret0
 }
 
-// Listeners indicates an expected call of Listeners.
-func (mr *MockStatusTrackingGatewayMockRecorder) Listeners() *gomock.Call {
+// Remove indicates an expected call of Remove.
+func (mr *MockStatusTrackingGatewayMockRecorder) Remove(ctx, id interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Listeners", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Listeners))
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Remove", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Remove), ctx, id)
 }
 
-// Meta mocks base method.
-func (m *MockStatusTrackingGateway) Meta() map[string]string {
+// Resolve mocks base method.
+func (m *MockStatusTrackingGateway) Resolve() core.ResolvedGateway {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Meta")
-	ret0, _ := ret[0].(map[string]string)
+	ret := m.ctrl.Call(m, "Resolve")
+	ret0, _ := ret[0].(core.ResolvedGateway)
 	return ret0
 }
 
-// Meta indicates an expected call of Meta.
-func (mr *MockStatusTrackingGatewayMockRecorder) Meta() *gomock.Call {
+// Resolve indicates an expected call of Resolve.
+func (mr *MockStatusTrackingGatewayMockRecorder) Resolve() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Meta", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Meta))
-}
-
-// ShouldBind mocks base method.
-func (m *MockStatusTrackingGateway) ShouldBind(route store.Route) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldBind", route)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldBind indicates an expected call of ShouldBind.
-func (mr *MockStatusTrackingGatewayMockRecorder) ShouldBind(route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldBind", reflect.TypeOf((*MockStatusTrackingGateway)(nil).ShouldBind), route)
-}
-
-// ShouldUpdate mocks base method.
-func (m *MockStatusTrackingGateway) ShouldUpdate(other store.Gateway) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldUpdate", other)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldUpdate indicates an expected call of ShouldUpdate.
-func (mr *MockStatusTrackingGatewayMockRecorder) ShouldUpdate(other interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldUpdate", reflect.TypeOf((*MockStatusTrackingGateway)(nil).ShouldUpdate), other)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Resolve", reflect.TypeOf((*MockStatusTrackingGateway)(nil).Resolve))
 }
 
 // TrackSync mocks base method.
@@ -143,6 +144,35 @@ func (m *MockGateway) EXPECT() *MockGatewayMockRecorder {
 	return m.recorder
 }
 
+// Bind mocks base method.
+func (m *MockGateway) Bind(ctx context.Context, route store.Route) []string {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "Bind", ctx, route)
+	ret0, _ := ret[0].([]string)
+	return ret0
+}
+
+// Bind indicates an expected call of Bind.
+func (mr *MockGatewayMockRecorder) Bind(ctx, route interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Bind", reflect.TypeOf((*MockGateway)(nil).Bind), ctx, route)
+}
+
+// CanFetchSecrets mocks base method.
+func (m *MockGateway) CanFetchSecrets(ctx context.Context, secrets []string) (bool, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "CanFetchSecrets", ctx, secrets)
+	ret0, _ := ret[0].(bool)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// CanFetchSecrets indicates an expected call of CanFetchSecrets.
+func (mr *MockGatewayMockRecorder) CanFetchSecrets(ctx, secrets interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanFetchSecrets", reflect.TypeOf((*MockGateway)(nil).CanFetchSecrets), ctx, secrets)
+}
+
 // ID mocks base method.
 func (m *MockGateway) ID() core.GatewayID {
 	m.ctrl.T.Helper()
@@ -157,244 +187,32 @@ func (mr *MockGatewayMockRecorder) ID() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockGateway)(nil).ID))
 }
 
-// Listeners mocks base method.
-func (m *MockGateway) Listeners() []store.Listener {
+// Remove mocks base method.
+func (m *MockGateway) Remove(ctx context.Context, id string) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Listeners")
-	ret0, _ := ret[0].([]store.Listener)
+	ret := m.ctrl.Call(m, "Remove", ctx, id)
+	ret0, _ := ret[0].(error)
 	return ret0
 }
 
-// Listeners indicates an expected call of Listeners.
-func (mr *MockGatewayMockRecorder) Listeners() *gomock.Call {
+// Remove indicates an expected call of Remove.
+func (mr *MockGatewayMockRecorder) Remove(ctx, id interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Listeners", reflect.TypeOf((*MockGateway)(nil).Listeners))
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Remove", reflect.TypeOf((*MockGateway)(nil).Remove), ctx, id)
 }
 
-// Meta mocks base method.
-func (m *MockGateway) Meta() map[string]string {
+// Resolve mocks base method.
+func (m *MockGateway) Resolve() core.ResolvedGateway {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Meta")
-	ret0, _ := ret[0].(map[string]string)
+	ret := m.ctrl.Call(m, "Resolve")
+	ret0, _ := ret[0].(core.ResolvedGateway)
 	return ret0
 }
 
-// Meta indicates an expected call of Meta.
-func (mr *MockGatewayMockRecorder) Meta() *gomock.Call {
+// Resolve indicates an expected call of Resolve.
+func (mr *MockGatewayMockRecorder) Resolve() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Meta", reflect.TypeOf((*MockGateway)(nil).Meta))
-}
-
-// ShouldBind mocks base method.
-func (m *MockGateway) ShouldBind(route store.Route) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldBind", route)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldBind indicates an expected call of ShouldBind.
-func (mr *MockGatewayMockRecorder) ShouldBind(route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldBind", reflect.TypeOf((*MockGateway)(nil).ShouldBind), route)
-}
-
-// ShouldUpdate mocks base method.
-func (m *MockGateway) ShouldUpdate(other store.Gateway) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ShouldUpdate", other)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// ShouldUpdate indicates an expected call of ShouldUpdate.
-func (mr *MockGatewayMockRecorder) ShouldUpdate(other interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ShouldUpdate", reflect.TypeOf((*MockGateway)(nil).ShouldUpdate), other)
-}
-
-// MockRouteTrackingListener is a mock of RouteTrackingListener interface.
-type MockRouteTrackingListener struct {
-	ctrl     *gomock.Controller
-	recorder *MockRouteTrackingListenerMockRecorder
-}
-
-// MockRouteTrackingListenerMockRecorder is the mock recorder for MockRouteTrackingListener.
-type MockRouteTrackingListenerMockRecorder struct {
-	mock *MockRouteTrackingListener
-}
-
-// NewMockRouteTrackingListener creates a new mock instance.
-func NewMockRouteTrackingListener(ctrl *gomock.Controller) *MockRouteTrackingListener {
-	mock := &MockRouteTrackingListener{ctrl: ctrl}
-	mock.recorder = &MockRouteTrackingListenerMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockRouteTrackingListener) EXPECT() *MockRouteTrackingListenerMockRecorder {
-	return m.recorder
-}
-
-// CanBind mocks base method.
-func (m *MockRouteTrackingListener) CanBind(ctx context.Context, route store.Route) (bool, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CanBind", ctx, route)
-	ret0, _ := ret[0].(bool)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// CanBind indicates an expected call of CanBind.
-func (mr *MockRouteTrackingListenerMockRecorder) CanBind(ctx, route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanBind", reflect.TypeOf((*MockRouteTrackingListener)(nil).CanBind), ctx, route)
-}
-
-// Config mocks base method.
-func (m *MockRouteTrackingListener) Config() store.ListenerConfig {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Config")
-	ret0, _ := ret[0].(store.ListenerConfig)
-	return ret0
-}
-
-// Config indicates an expected call of Config.
-func (mr *MockRouteTrackingListenerMockRecorder) Config() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Config", reflect.TypeOf((*MockRouteTrackingListener)(nil).Config))
-}
-
-// ID mocks base method.
-func (m *MockRouteTrackingListener) ID() string {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ID")
-	ret0, _ := ret[0].(string)
-	return ret0
-}
-
-// ID indicates an expected call of ID.
-func (mr *MockRouteTrackingListenerMockRecorder) ID() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockRouteTrackingListener)(nil).ID))
-}
-
-// IsValid mocks base method.
-func (m *MockRouteTrackingListener) IsValid() bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "IsValid")
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// IsValid indicates an expected call of IsValid.
-func (mr *MockRouteTrackingListenerMockRecorder) IsValid() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsValid", reflect.TypeOf((*MockRouteTrackingListener)(nil).IsValid))
-}
-
-// OnRouteAdded mocks base method.
-func (m *MockRouteTrackingListener) OnRouteAdded(route store.Route) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnRouteAdded", route)
-}
-
-// OnRouteAdded indicates an expected call of OnRouteAdded.
-func (mr *MockRouteTrackingListenerMockRecorder) OnRouteAdded(route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnRouteAdded", reflect.TypeOf((*MockRouteTrackingListener)(nil).OnRouteAdded), route)
-}
-
-// OnRouteRemoved mocks base method.
-func (m *MockRouteTrackingListener) OnRouteRemoved(id string) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnRouteRemoved", id)
-}
-
-// OnRouteRemoved indicates an expected call of OnRouteRemoved.
-func (mr *MockRouteTrackingListenerMockRecorder) OnRouteRemoved(id interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnRouteRemoved", reflect.TypeOf((*MockRouteTrackingListener)(nil).OnRouteRemoved), id)
-}
-
-// MockListener is a mock of Listener interface.
-type MockListener struct {
-	ctrl     *gomock.Controller
-	recorder *MockListenerMockRecorder
-}
-
-// MockListenerMockRecorder is the mock recorder for MockListener.
-type MockListenerMockRecorder struct {
-	mock *MockListener
-}
-
-// NewMockListener creates a new mock instance.
-func NewMockListener(ctrl *gomock.Controller) *MockListener {
-	mock := &MockListener{ctrl: ctrl}
-	mock.recorder = &MockListenerMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockListener) EXPECT() *MockListenerMockRecorder {
-	return m.recorder
-}
-
-// CanBind mocks base method.
-func (m *MockListener) CanBind(ctx context.Context, route store.Route) (bool, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CanBind", ctx, route)
-	ret0, _ := ret[0].(bool)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// CanBind indicates an expected call of CanBind.
-func (mr *MockListenerMockRecorder) CanBind(ctx, route interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanBind", reflect.TypeOf((*MockListener)(nil).CanBind), ctx, route)
-}
-
-// Config mocks base method.
-func (m *MockListener) Config() store.ListenerConfig {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Config")
-	ret0, _ := ret[0].(store.ListenerConfig)
-	return ret0
-}
-
-// Config indicates an expected call of Config.
-func (mr *MockListenerMockRecorder) Config() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Config", reflect.TypeOf((*MockListener)(nil).Config))
-}
-
-// ID mocks base method.
-func (m *MockListener) ID() string {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ID")
-	ret0, _ := ret[0].(string)
-	return ret0
-}
-
-// ID indicates an expected call of ID.
-func (mr *MockListenerMockRecorder) ID() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockListener)(nil).ID))
-}
-
-// IsValid mocks base method.
-func (m *MockListener) IsValid() bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "IsValid")
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// IsValid indicates an expected call of IsValid.
-func (mr *MockListenerMockRecorder) IsValid() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "IsValid", reflect.TypeOf((*MockListener)(nil).IsValid))
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Resolve", reflect.TypeOf((*MockGateway)(nil).Resolve))
 }
 
 // MockStatusTrackingRoute is a mock of StatusTrackingRoute interface.
@@ -434,30 +252,6 @@ func (mr *MockStatusTrackingRouteMockRecorder) ID() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockStatusTrackingRoute)(nil).ID))
 }
 
-// OnBindFailed mocks base method.
-func (m *MockStatusTrackingRoute) OnBindFailed(err error, gateway store.Gateway) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnBindFailed", err, gateway)
-}
-
-// OnBindFailed indicates an expected call of OnBindFailed.
-func (mr *MockStatusTrackingRouteMockRecorder) OnBindFailed(err, gateway interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnBindFailed", reflect.TypeOf((*MockStatusTrackingRoute)(nil).OnBindFailed), err, gateway)
-}
-
-// OnBound mocks base method.
-func (m *MockStatusTrackingRoute) OnBound(gateway store.Gateway) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "OnBound", gateway)
-}
-
-// OnBound indicates an expected call of OnBound.
-func (mr *MockStatusTrackingRouteMockRecorder) OnBound(gateway interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnBound", reflect.TypeOf((*MockStatusTrackingRoute)(nil).OnBound), gateway)
-}
-
 // OnGatewayRemoved mocks base method.
 func (m *MockStatusTrackingRoute) OnGatewayRemoved(gateway store.Gateway) {
 	m.ctrl.T.Helper()
@@ -468,20 +262,6 @@ func (m *MockStatusTrackingRoute) OnGatewayRemoved(gateway store.Gateway) {
 func (mr *MockStatusTrackingRouteMockRecorder) OnGatewayRemoved(gateway interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "OnGatewayRemoved", reflect.TypeOf((*MockStatusTrackingRoute)(nil).OnGatewayRemoved), gateway)
-}
-
-// Resolve mocks base method.
-func (m *MockStatusTrackingRoute) Resolve(listener store.Listener) core.ResolvedRoute {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Resolve", listener)
-	ret0, _ := ret[0].(core.ResolvedRoute)
-	return ret0
-}
-
-// Resolve indicates an expected call of Resolve.
-func (mr *MockStatusTrackingRouteMockRecorder) Resolve(listener interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Resolve", reflect.TypeOf((*MockStatusTrackingRoute)(nil).Resolve), listener)
 }
 
 // SyncStatus mocks base method.
@@ -535,20 +315,6 @@ func (mr *MockRouteMockRecorder) ID() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ID", reflect.TypeOf((*MockRoute)(nil).ID))
 }
 
-// Resolve mocks base method.
-func (m *MockRoute) Resolve(listener store.Listener) core.ResolvedRoute {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Resolve", listener)
-	ret0, _ := ret[0].(core.ResolvedRoute)
-	return ret0
-}
-
-// Resolve indicates an expected call of Resolve.
-func (mr *MockRouteMockRecorder) Resolve(listener interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Resolve", reflect.TypeOf((*MockRoute)(nil).Resolve), listener)
-}
-
 // MockStore is a mock of Store interface.
 type MockStore struct {
 	ctrl     *gomock.Controller
@@ -570,21 +336,6 @@ func NewMockStore(ctrl *gomock.Controller) *MockStore {
 // EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockStore) EXPECT() *MockStoreMockRecorder {
 	return m.recorder
-}
-
-// CanFetchSecrets mocks base method.
-func (m *MockStore) CanFetchSecrets(ctx context.Context, id core.GatewayID, secrets []string) (bool, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CanFetchSecrets", ctx, id, secrets)
-	ret0, _ := ret[0].(bool)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// CanFetchSecrets indicates an expected call of CanFetchSecrets.
-func (mr *MockStoreMockRecorder) CanFetchSecrets(ctx, id, secrets interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CanFetchSecrets", reflect.TypeOf((*MockStore)(nil).CanFetchSecrets), ctx, id, secrets)
 }
 
 // DeleteGateway mocks base method.
@@ -615,19 +366,19 @@ func (mr *MockStoreMockRecorder) DeleteRoute(ctx, id interface{}) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteRoute", reflect.TypeOf((*MockStore)(nil).DeleteRoute), ctx, id)
 }
 
-// GatewayExists mocks base method.
-func (m *MockStore) GatewayExists(ctx context.Context, id core.GatewayID) (bool, error) {
+// GetGateway mocks base method.
+func (m *MockStore) GetGateway(ctx context.Context, id core.GatewayID) (store.Gateway, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GatewayExists", ctx, id)
-	ret0, _ := ret[0].(bool)
+	ret := m.ctrl.Call(m, "GetGateway", ctx, id)
+	ret0, _ := ret[0].(store.Gateway)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
-// GatewayExists indicates an expected call of GatewayExists.
-func (mr *MockStoreMockRecorder) GatewayExists(ctx, id interface{}) *gomock.Call {
+// GetGateway indicates an expected call of GetGateway.
+func (mr *MockStoreMockRecorder) GetGateway(ctx, id interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GatewayExists", reflect.TypeOf((*MockStore)(nil).GatewayExists), ctx, id)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetGateway", reflect.TypeOf((*MockStore)(nil).GetGateway), ctx, id)
 }
 
 // Sync mocks base method.
